replace rgb_map_mutate by map_mutate to improve performance
The new algorithm only uses one function that return a pixel instead of
3. This helps optimizing by reducing function calls.
