+++ src/hhimg/hhimg.h
@@ -10,11 +10,12 @@
 #include "./algorithm/minus.h"
 #include "./algorithm/convolution.h"
 #include "./algorithm/rgb_map.h"
-#include "./algorithm/rgb_map_mutate.h"
+#include "./algorithm/map_mutate.h"
 #include "./tools/perf_recorder.h"
 #include "./tools/algorithm_operators.h"
 #include "./tools/utils.h"
 #include "./abstract/abstract_algorithm.h"
+#include "./abstract/abstract_hh_algorithm.h"
 #include "./abstract/abstract_tile_algorithm.h"
 #include "./abstract/abstract_pair_tile_algorithm.h"
 #include "./abstract/abstract_image_factory.h"
